Update tuple feature to recent changes
