+++ src/lib.rs
@@ -1,8 +1,8 @@
-#[cfg(feature="tuple")]
-extern crate tuple;
-
-mod stepper;
 mod ode;
+mod stepper;
+
+#[cfg(feature = "tuple")]
+mod tuples;
 
 // Re-exports
 pub use ode::Ode;
