+++ src/ode.rs
@@ -1,9 +1,3 @@
-#[cfg(feature = "tuple")]
-mod tuple_impls;
-
-#[cfg(feature = "tuple")]
-pub use self::tuple_impls::*;
-
 pub trait Ode {
     type State: Clone;
